refactoring classes to separate files (closes #113, #111)
