+++ lib/error/PodioAuthorizationError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioAuthorizationError extends PodioError
+{
+}
