+++ lib/error/PodioBadRequestError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioBadRequestError extends PodioError
+{
+}
