+++ lib/error/PodioConflictError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioConflictError extends PodioError
+{
+}
