+++ lib/error/PodioConnectionError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioConnectionError extends Exception
+{
+}
