+++ lib/error/PodioDataIntegrityError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioDataIntegrityError extends Exception
+{
+}
