+++ lib/error/PodioForbiddenError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioForbiddenError extends PodioError
+{
+}
