+++ lib/error/PodioGoneError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioGoneError extends PodioError
+{
+}
