+++ lib/error/PodioInvalidGrantError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioInvalidGrantError extends PodioError
+{
+}
