+++ lib/error/PodioMissingRelationshipError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioMissingRelationshipError extends PodioError
+{
+}
