+++ lib/error/PodioNotFoundError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioNotFoundError extends PodioError
+{
+}
