+++ lib/error/PodioRateLimitError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioRateLimitError extends PodioError
+{
+}
