+++ lib/error/PodioServerError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioServerError extends PodioError
+{
+}
