+++ lib/error/PodioUnavailableError.php
@@ -1,0 +1,5 @@
+<?php
+
+class PodioUnavailableError extends PodioError
+{
+}
