+++ models/field/PodioEmailItemField.php
@@ -1,0 +1,9 @@
+<?php
+
+/**
+ * email field
+ */
+class PodioEmailItemField extends PodioPhoneOrEmailItemField
+{
+
+}
