+++ models/field/PodioFileItemField.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * File field
+ */
+class PodioFileItemField extends PodioAssetItemField
+{
+}
