+++ models/field/PodioImageItemField.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * Image field
+ */
+class PodioImageItemField extends PodioAssetItemField
+{
+}
