+++ models/field/PodioPhoneItemField.php
@@ -1,0 +1,9 @@
+<?php
+
+/**
+ * phone field
+ */
+class PodioPhoneItemField extends PodioPhoneOrEmailItemField
+{
+
+}
